move constants out of integeration tests
